Move some code out from kernel_cache
